+++ api/extjs/new-api/components/tab.js
@@ -6,21 +6,6 @@ const compName = 'Tab';
 
 const actions = {
   compName,
-
-  async click(tEQ, idForLog, enableLog) {
-    return gIn.wrap({
-      msg: `${compName}  "${tEQ}": click ... `,
-      enableLog,
-      act: async () => {
-        const id = await queryCmpId(
-          tEQ,
-          idForLog,
-          false
-        );
-        await gT.s.uA.clickById(id, false);
-      },
-    });
-  },
 };
 
 const checks = {