+++ types/ext-js/ext-js-components/tab.d.ts
@@ -2,10 +2,6 @@ import {ElementIdForLog, EnableLog, Teq} from '../common';
 import {ComponentActions, ComponentChecks, ComponentLogs} from './component';
 
 interface TabActions extends ComponentActions {
-  /**
-   * Left mouse button click.
-   */
-  click(tEQ: Teq, idForLog: ElementIdForLog, enableLog: EnableLog): Promise<undefined>;
 }
 
 interface TabChecks extends ComponentChecks {